+++ tsconfig.app.json
@@ -26,6 +26,11 @@
   "include": ["src",
               ".storybook/vitest.setup.ts",
               ".storybook/**/*.ts",
-              "vitest.workspace.ts"],
+              "vitest.workspace.ts",
+              "stories/**/*.ts",
+              "stories/**/*.tsx", 
+              ".storybook/preview.tsx",
+              ".storybook/**/*.ts",
+              ".storybook/**/*.tsx"],
   "exclude": ["node_modules", "dist"]
 }
